Moving tests over to the Components folder. This is more in line with the jest styling conventions! Makes it easier to find tests.
Also updating the depedencies with a few things that I think we're installed globally before. This should help with people contributing in the future :)

I had to update a few jsx-a11y rules to make sure that we can commit without eslint throwing errors. Let me know if this doesn't work but maybe a diferent solution can be had.